- Summary: 1. Bugzilla Bug 120922 "Dirty Page" behavior in XML Report Designer editor needs improved 2. Bugzilla Bug 118916	BIRT JavaScript objects are not in alpha order in expression builder
- Bugzilla Bug (s) Resolved:Bugzilla Bug 118916,120922

- Description:
1.Add new message dialog to let user choice to save,doesn't save or cancel.
The only different with the suggesion from Paul is, if user select no the chages will not be discarded automatically. Becuase if we do auto discarding, we need reload the file and rebuild all graphical designer. That can be a potential performace issue.
So if user select no the page will switch to graphical editor. The user can simplely click the save button to discard all source changes.
2. Sort classes,methods, and members in alphabitical order in expression builder.

- Tests Description: manual

- Files Edited:

- Files Added:

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation: