- Summary: fix 136893 Preview Report repeats first line for all entries
- Bugzilla Bug (s) Resolved:
136893
- Description:
When drag a dataset column into layout:
Search the target container, if container has the same dataset, add the column binding if it does not exist in the container. If the container's dataset is not the dragged dataset column's dataset, column binding will be added to the new dataitem, and set dataitem's dataset with the dragged dataset column's dataset.


- Tests Description:
Manual test.
- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Edited:

- Files Added: