-Summary: [Linux-WS]Width of the selection tital
- Bugzilla Bug (s) Resolved:134999

- Description:
 It works now:
I added a method to calculate the length in pixels of given string,and set the length to the label control, and it can properly display now.
But I'm not sure whether it works well in non-English environment, for I can't test it here. So can you help me test it, Yi? Thank you indeed.
- Tests Description: Manual

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:

- Files Edited: