-Summary:[regression]Warning dialog when select a key/value from read-only properties file.
- Bugzilla Bug (s) Resolved:145266

- Description:
 I've fixed another bug 130499,whose theme is Offer some warning if add new
key/value to a read-only resource file in March this year.
Afrer Ok button is pressed, it'll save key/values to property file even if the
key/value list is not changed. And before saving file, I checked whether the
file is read only or not. If it's read only, then an error dialog is show.

The above is the reason why this bug happens.

If key/value list is unchanged, I won't save this file.

- Tests Description: Manual

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:

- Files Edited: