- Summary: [Context Sensitive] Can't link to correct topic the first time(main and 2.1.1 branch).
- Bugzilla Bug (s) Resolved:
 148754

- Description:
Each Help topic binds to a control.
When the wizard shows the page, its control is not focused on, instead its parent is. So when you click help mark(?), eclipse will look for its parent's help content, which is not expected.
Now, I'ts fixed.

Check in to main and 2.1.1 brunch.

- Tests Description:
Manual test

- Files Edited:

@changelist

- Files Added:

@addlist

- Files Deleted:

@deletelist

- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: