Summary Fix Bugzilla bug 151353-UI improvement of the hyperlink builder
Bugzilla Bug (s) Resolved 151353

Description
Add a value of "---" into the bookmark and TOC selection list, and always display it in the drop down list. If user selects "---", set the link not include bookmark and TOC.

Tests Description Manual test and code review.

Notes to Build Team

Notes to Developers

Notes to QA

Notes to Documentation