- Summary: NPE when opening attached report design
- Bugzilla Bug (s) Resolved:
 166417

- Description:
When opening the attached report design(https://bugs.eclipse.org/bugs/attachment.cgi?id=54837) in 2.1.2, it throws an exception,but works well in 2.2.0.  The reason is that we used a ModuleHandle before checking whether it's null or not. The problem occurs after we fixed another bug, and the bug has been fixed in 2.2.0, but not in 2.1.2.
Fixed.


- Tests Description:
Manual test

- Files Edited:



- Files Added:



- Files Deleted:



- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: