- Summary: Cascading parameter dialog - Adding new parameter
- Bugzilla Bug (s) Resolved:
 170998

- Description:
User is able to give parameter a name in Parameters Name field.
There is an option for the user to either give a name first or choose a
data set.
After selecting the data set, related properties are visible under
properties box with Name as read only. To change the parameter name user can
go to parameters field and change the name.
Fixed.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: