Fixing bug 183490.
auto table layout:
The master page border will be extended as the content grows (similar to table
column width). This is the main cause for issue 2 because the design layout
editor is limited by the page size. But in HTML output, it won¡¯t honor the page
size.
Still use the same algorithm for table width and column width calculation.
When adding a new report item into a table/grid cell:
text/label item, it should wrap the text/label using the same algorithm as in
HTML if the content is longer than the column width. If the content can¡¯t be
wrapped, it will extend the column to display the full text.
image item/chart,  extend the column to display the full image item/chart
container item (grid, table, list, x-tab):   if the container item¡¯s size is
larger than the cell¡¯s width, extend the column to display the item. The
container size should be calculated in the same way as the table width in Wei¡¯s
algorithm.

This enhancement makes the layout of designer be similar with engine's html
output.