- Summary:Fix Bugzilla 187496-Remove Total functions from default expression builder
- Bugzilla Bug (s) Resolved: 187496

- Description:
This bug is duplicate with 187178, and is fixed and checked in before. Then the checkin is required roll back. The checkin of DEUtil is roll back, checkins on other files are not tagged in RC0.

Now, check in the fix on DEUtil for this bug again. And all the fixes will be tagged in RC2.

- Tests Description: Manual test and code review

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Edited:


- Files Added:


- Files Deleted: