- Summary: Cross Tab: Filter dialog layout changes
- Bugzilla Bug (s) Resolved:
206065

- Description:
Update Crosstab Filter Condition Dialog following the 3 attached  mockups.
In addition, I implemented what Rima said in comment 3:For operators other than 'IN',move the second box ( column/row member value ) up and not leave a gap in between.
But I didn't remove "Add" button when the operator is In/Not In. That's because users not only can select a value from list, but they also can input value through keyboard or expression builder dialog, and user will add this value to the list by clicking "Add" button, or they have no way to do that.
Fixed.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: