- Summary: Error occurs when change the value of a cascading parameter.
- Bugzilla Bug (s) Resolved:
210622

- Description:
In step 6, when you change the value from CITY to CUSTOMERNUMBER, the data type also updates.The previous value is string,  and now it's integer.  For parameters with different data types have different format choice, and When setting Uppercase to string type, it's OK, but invalid for  integer type.
So when data type is changed, if the format type is invalid for the new type, then set it to "Unformatted".


- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: