- Summary: Can not Remove Default while no default is required [0705]
- Bugzilla Bug (s) Resolved:
224520

- Description:
We cannot differ the 2 conditions: 1.default value is not set; 2.default value is set to null.
So this bug cannot be fixed until model provide more API.

It has no correlation between "Is required" and "Default value must be set", and default value can be null even if "Is required" checkbox is checked.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: