- Summary: In Chinese locale, parameter value in drill-through is not stored correctly.(GUI part)
- Bugzilla Bug (s) Resolved:
236585

- Description:
Encoding inconsistency in expression builder.
GUI part is fixed, and now you'll see the result is right in step 6.
But in step 7, user may find that actual action is not expected.
I find model saves the value correctly, so reassign this bug to Engine to fix the rest.

- Tests Description  :
Manual test


- Files Edited:



- Files Added:



- Files Deleted:



- Notes to Build Team:


- Notes to Developers:


- Notes to QA:


-  Notes to Documentation: