- Summary: fix 244916 Warning message is incorrect of inserting a MODE Function aggregation
- Bugzilla Bug (s) Resolved:244916

- Description:
This is caused by 'any' data type is deprecated in model but still used in DTE, if the DTE recommended datatype is 'any', we can not get the dispaly name from model to compare with use selected datatype name, thus popup the warning. Resolved this by ignoring the 'any' data type.

- Tests Description  :
Manual test

- Files Edited:


- Files Added:


- Files Deleted:


- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

-  Notes to Documentation: