- Summary: [RCP] No help dialogue pops up when click "?", and exception is thrown out when open "Help" View [00]
- Bugzilla Bug (s) Resolved:
255302

- Description:
About issue 1. Added help Id for this page.
About issue 2. I can't reproduce it, it works without any exception here. Even you can reproduce it, it's not a bug of BIRT, for help view is managed by eclipse platform, not us.  So, I reassign this bug to build team to check if there's a build issue. If the build is OK, please set it to FIXED.
Checked into head and 2.3.2 branch.

- Tests Description  :
Manual test


- Files Edited:



- Files Added:




- Files Deleted:





- Notes to Build Team:



- Notes to Developers:



- Notes to QA:



-  Notes to Documentation: